+++ fsutil/fs_darwin.go
@@ -61,3 +61,9 @@ func GetTimestamps(path string) (int64, int64, int64, error) {
 		int64(stat.Ctimespec.Sec),
 		nil
 }
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+func isEmptyDirent(n int) bool {
+	return n == 0x30
+}