+++ system/user_bsd.go
@@ -1,6 +1,6 @@
-// +build linux
+// +build darwin freebsd
 
-package fsutil
+package system
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -12,19 +12,15 @@ package fsutil
 import (
 	"syscall"
 	"time"
-
-	PATH "pkg.re/essentialkaos/ek.v8/path"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// GetTimes return time of access, modification and creation at once
-func GetTimes(path string) (time.Time, time.Time, time.Time, error) {
+// getTimes is copy of fsutil.GetTimes
+func getTimes(path string) (time.Time, time.Time, time.Time, error) {
 	if path == "" {
 		return time.Time{}, time.Time{}, time.Time{}, ErrEmptyPath
 	}
-
-	path = PATH.Clean(path)
 
 	var stat = &syscall.Stat_t{}
 
@@ -34,30 +30,8 @@ func GetTimes(path string) (time.Time, time.Time, time.Time, error) {
 		return time.Time{}, time.Time{}, time.Time{}, err
 	}
 
-	return time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec)),
-		time.Unix(int64(stat.Mtim.Sec), int64(stat.Mtim.Nsec)),
-		time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec)),
-		nil
-}
-
-// GetTimestamps return time of access, modification and creation at once as linux timestamp
-func GetTimestamps(path string) (int64, int64, int64, error) {
-	if path == "" {
-		return -1, -1, -1, ErrEmptyPath
-	}
-
-	path = PATH.Clean(path)
-
-	var stat = &syscall.Stat_t{}
-
-	err := syscall.Stat(path, stat)
-
-	if err != nil {
-		return -1, -1, -1, err
-	}
-
-	return int64(stat.Atim.Sec),
-		int64(stat.Mtim.Sec),
-		int64(stat.Ctim.Sec),
+	return time.Unix(int64(stat.Atimespec.Sec), int64(stat.Atimespec.Nsec)),
+		time.Unix(int64(stat.Mtimespec.Sec), int64(stat.Mtimespec.Nsec)),
+		time.Unix(int64(stat.Ctimespec.Sec), int64(stat.Ctimespec.Nsec)),
 		nil
 }